requirements/base.txt
Django==1.8.5
Unipath==1.1
requests==2.7.0
django-haystack==2.4.0
pysolr==3.3.2
